+++ dwertheimer.EventAutomations/src/NPTimeblocking.js
@@ -11,6 +11,7 @@
 import { differenceInCalendarDays, endOfDay, startOfDay, eachMinuteOfInterval, formatISO9075, addMinutes } from 'date-fns'
 import { getTimedEntries, keepTodayPortionOnly } from '../../helpers/calendar'
 import { getEventsForDay, writeTimeBlocksToCalendar } from '../../helpers/NPCalendar'
+import { getParagraphBlock } from '../../jgclark.Filer/src/fileItems'
 import {
   getDateStringFromCalendarFilename,
   type HourMinObj,
@@ -29,7 +30,7 @@ import { sortListBy } from '../../helpers/sorting'
 import { showMessage, chooseOption } from '../../helpers/userInput'
 import { isTimeBlockLine, getTimeBlockString } from '../../helpers/timeblocks'
 import { calcSmartPrependPoint } from '../../helpers/paragraph'
-import { logAllPropertyNames, getAllPropertyNames, JSP, clo } from '../../helpers/dev'
+import { logAllPropertyNames, getAllPropertyNames, JSP, clo, log } from '../../helpers/dev'
 import {
   attachTimeblockTag,
   blockOutEvents,
@@ -51,7 +52,7 @@ import { getTimeBlockingDefaults, validateTimeBlockConfig } from './config'
 import type { IntervalMap, TimeBlockDefaults, PartialCalendarItem, EditorOrNote } from './timeblocking-flow-types'
 import { checkNumber, checkObj, checkString, checkWithDefault } from '../../helpers/checkType'
 import { catchError } from 'rxjs/operators'
-
+import pluginJson from '../plugin.json'
 const PLUGIN_ID = 'autoTimeBlocking'
 
 /* TCalendarItem is a type for the calendar items:
@@ -96,12 +97,13 @@ const editorIsOpenToToday = () => {
 }
 
 async function insertContentUnderHeading(destNote: TNote, headingToFind: string, parasAsText: string) {
+  // console.log(`insertContentUnderHeading(${headingToFind}, ${parasAsText})`)
   const topOfNote = destNote.type === 'Calendar' ? 0 : 1
   let insertionIndex = topOfNote // top of note by default
   //   console.log(`insertionIndex:${insertionIndex}`)
   for (let i = 0; i < destNote.paragraphs.length; i++) {
     const p = destNote.paragraphs[i]
-    if (p.content.trim() === headingToFind && p.type === 'title') {
+    if (p.content.trim().includes(headingToFind) && p.type === 'title') {
       insertionIndex = i + 1
       break
     }
@@ -133,7 +135,8 @@ export async function deleteParagraphsContainingString(destNote: TNote, timeBloc
  * backlinks[0].subItems[0] =JSLog: {"type":"open","content":"scheduled for 10/4 using app >today","rawContent":"* scheduled for 10/4 using app
  * ","prefix":"* ","contentRange":{},"lineIndex":2,"date":"2021-11-07T07:00:00.000Z","heading":"_Testing scheduled sweeping","headingRange":{},"headingLevel":1,"isRecurring":0,"indents":0,"filename":"zDELETEME/Test scheduled.md","noteType":"Notes","linkedNoteTitles":[],"subItems":[]}
  */
-function getTodaysReferences(pNote: TNote | null = null, config): Array<TParagraph> {
+function getTodaysReferences(pNote: TNote | null = null, config: { [key: string]: any }): Array<TParagraph> {
+  log(pluginJson, `getTodaysReferences starting`)
   const note = pNote || Editor.note
   if (note == null) {
     console.log(`timeblocking could not open Note`)
@@ -141,31 +144,49 @@ function getTodaysReferences(pNote: TNote | null = null, config): Array<TParagra
   }
 
   const backlinks: Array<TParagraph> = [...note.backlinks] // an array of notes which link to this note
+  log(pluginJson, `backlinks.length:${backlinks.length}`)
 
+  // clo(backlinks, 'getTodaysReferences: backlinks')
   let todayParas = []
   backlinks.forEach((link, i) => {
-    const subItems = link.children()
+    // $FlowIgnore Flow(prop-missing) -- subItems is not in Flow defs but is real
+    const subItems = link.subItems
     subItems.forEach((subItem, j) => {
       subItem.title = link.content.replace('.md', '').replace('.txt', '')
       todayParas.push(subItem)
     })
   })
-  // FIX CLO to work with backlinks
 
-  clo(todayParas, 'todayParas')
-  findTodosInNote(note).forEach((link, i) => {
-    clo(link, `findTodosInNote[${i}]`)
-  })
-  todayParas = [...todayParas, ...findTodosInNote(note)]
+  // clo(todayParas, 'todayParas')
+  if (config.createSyncedCopies) {
+    log(pluginJson, `getTodaysReferences: Cannot search today's note for >today items when config.createSyncedCopies is on, because could be recursive`)
+  } else {
+    findTodosInNote(note).forEach((link, i) => {
+      clo(link, `findTodosInNote[${i}]`)
+    })
+    todayParas = [...todayParas, ...findTodosInNote(note)]
+  }
   // console.log(`getTodaysReferences note.filename=${note.filename} backlinks.length=${backlinks.length} todayParas.length=${todayParas.length}`)
   return todayParas
 }
 
-async function insertItemsIntoNote(note, list, config) {
-  const { timeBlockHeading } = config
+async function insertItemsIntoNote(note: TNote, list: Array<string>, heading: string = '', shouldFold: boolean = false, config: { [string]: any }) {
   if (list && list.length > 0 && note) {
     // $FlowIgnore
-    await insertContentUnderHeading(note, timeBlockHeading, list.join('\n'))
+    await insertContentUnderHeading(note, heading, list.join('\n'))
+    // Fold the heading to hide the list
+    if (shouldFold && heading !== '') {
+      const thePara = note.paragraphs.find((p) => p.type == 'title' && p.content.includes(heading))
+      if (thePara) {
+        log(pluginJson, `insertItemsIntoNote: folding "${heading}"`)
+        thePara.content = `${String(heading)} …`
+        note.updateParagraph(thePara)
+        log(pluginJson, `insertItemsIntoNote: thePara.content "${thePara.content}"`)
+        note.content = note.content //FIXME: hoping for an API to do this so we don't have to force a redraw so it will fold the heading
+      } else {
+        log(pluginJson, `insertItemsIntoNote could not find heading: ${heading}`)
+      }
+    }
   } else {
     if (!config.passBackResults) {
       await showMessage('No work hours left. Check config/presents.')
@@ -235,6 +256,28 @@ export async function deleteCalendarEventsWithTag(tag: string, dateStr: string):
   }
 }
 
+export async function deleteEntireBlock(note: TNote | Editor, para: TParagraph, useExtendedBlockDefinition: boolean = false, keepTitle: boolean = true) {
+  const paraBlock: Array<TParagraph> = getParagraphBlock(note, para.lineIndex)
+  log(pluginJson, `deleteEntireBlock: Removing ${paraBlock.length} items under ${para.content}`)
+  keepTitle ? paraBlock.shift() : null
+  if (paraBlock.length > 0) {
+    note.removeParagraphs(paraBlock) //seems to not work if it's a note, not Editor
+    // note.updateParagraphs(paraBlock)
+  }
+}
+
+export async function deleteBlockUnderTitle(note: TNote | Editor, title: string, useExtendedBlockDefinition: boolean = false) {
+  // log(pluginJson, `deleteBlockUnderTitle: ${note.title} Remove items under title: "${title}"`)
+  const para = note.paragraphs.find((p) => p.type == 'title' && p.content.includes(title))
+  let paraBlock = []
+  // clo(para, `deleteBlockUnderTitle para=`)
+  if (para && para.lineIndex != null) {
+    deleteEntireBlock(note, para, useExtendedBlockDefinition, true)
+  } else {
+    console.log(`deleteBlockUnderTitle: did not find title: "${title}"`)
+  }
+}
+
 type TEventConfig = {
   confirmEventCreation: boolean,
   processedTagName: string,
@@ -265,6 +308,55 @@ function getEventsConfig(atbConfig: { [string]: mixed }): TEventConfig {
   return eventsConfig
 }
 
+/**
+ * Make copies of all OPEN task references as Synced Lines and return them as an array of strings
+ * @param {*} allTodayParagraphs
+ * @returns array of strings with the sync codes attached
+ */
+export function getSyncedCopiesAsList(allTodayParagraphs: Array<TParagraph>): Array<string> {
+  // clo(allTodayParagraphs, 'getSyncedCopiesAsList::allTodayParagraphs')
+  let syncedLinesList = []
+  allTodayParagraphs.forEach((p) => {
+    if (p.type == 'open') {
+      p.note?.addBlockID(p)
+      p.note?.updateParagraph(p)
+      syncedLinesList.push(p.rawContent)
+    }
+  })
+  // timeBlockTextList.forEach((line) => {
+  //   syncedLinesList.push(line)
+  //   console.log(`getSyncedCopiesAsList::line=${line}:`)
+  //   const segments = line.split(' ')
+  //   if (segments.length > 2) {
+  //     let content = line.replace(`${segments[0]} ${segments[1]} `, '') // replace time at the front
+  //     content = content.replace(/( \[.*\]\(.*\).*)/, '') // replace pretty link
+  //     content = content.replace(/( \[\[.*\]\].*)/, '') // replace wiki link
+  //     console.log(`getSyncedCopiesAsList::content=${content}:`)
+  //     syncedLinesList.push(content)
+  //     // search in paragraphs
+  //   }
+  // const content = line.slice()
+  // p?.note?.addBlockID(p)
+  // p?.note?.updateParagraph(p)
+  // syncedLinesList.push(p.content)
+  // })
+  return syncedLinesList
+}
+
+function getFullParagraphsCorrespondingToSortList(paragraphs: Array<TParagraph>, sortList: Array<{ [string]: any }>): Array<TParagraph> {
+  let retP = []
+  if (sortList && paragraphs) {
+    retP = sortList.map((s) => {
+      const found = paragraphs.find((p, i) => {
+        return removeDateTagsAndToday(p.rawContent) === s.raw
+      })
+      return found
+    })
+  }
+  // $FlowIgnore
+  return retP ?? []
+}
+
 export async function createTimeBlocksForTodaysTasks(config: { [key: string]: mixed } = {}): Promise<?Array<string>> {
   // console.log(`Starting createTimeBlocksForTodaysTasks. Time is ${new Date().toLocaleTimeString()}`)
   // console.log(`config is: ${JSON.stringify(config, null, 2)}`)
@@ -280,45 +372,61 @@ export async function createTimeBlocksForTodaysTasks(config: { [key: string]: mi
     excludeTasksWithText,
   } = config
   const hypenatedDate = getTodaysDateHyphenated()
+  log(pluginJson, `createTimeBlocksForTodaysTasks hypenatedDate=${hypenatedDate}`)
   const date = getTodaysDateUnhyphenated()
+  log(pluginJson, `createTimeBlocksForTodaysTasks date=${date}`)
   const note = Editor // placeholder. we may pass a note in future revs
   const dateStr = Editor.filename ? getDateStringFromCalendarFilename(Editor.filename) : null
+  log(pluginJson, `createTimeBlocksForTodaysTasks dateStr=${dateStr}`)
   if (dateStr && dateStr === date) {
+    log(pluginJson, `createTimeBlocksForTodaysTasks dateStr=${dateStr} is today - we are inside`)
     const backlinkParas = getTodaysReferences(Editor.note, config)
-    // console.log(`Found ${backlinkParas.length} backlinks+today-note items (may include completed items)`)
+    console.log(`Found ${backlinkParas.length} backlinks+today-note items (may include completed items)`)
     let todosParagraphs: Array<TParagraph> = makeAllItemsTodos(backlinkParas) //some items may not be todos but we want to pretend they are and timeblock for them
     todosParagraphs = includeTasksWithText?.length ? includeTasksWithPatterns(todosParagraphs, includeTasksWithText) : todosParagraphs
-    // console.log(`After includeTasksWithText, ${todosParagraphs.length} potential items`)
+    console.log(`After includeTasksWithText, ${todosParagraphs.length} potential items`)
     todosParagraphs = excludeTasksWithText?.length ? excludeTasksWithPatterns(todosParagraphs, excludeTasksWithText) : todosParagraphs
-    // console.log(`After excludeTasksWithText, ${todosParagraphs.length} potential items`)
+    console.log(`After excludeTasksWithText, ${todosParagraphs.length} potential items`)
     const cleanTodayTodoParas = removeDateTagsFromArray(todosParagraphs)
-    // console.log(`After removeDateTagsFromArray, ${cleanTodayTodoParas.length} potential items`)
+    console.log(`After removeDateTagsFromArray, ${cleanTodayTodoParas.length} potential items`)
     const todosWithLinksMaybe = appendLinkIfNecessary(cleanTodayTodoParas, config)
-    // console.log(`After appendLinkIfNecessary, ${todosWithLinksMaybe?.length ?? 0} potential items (may include headings or completed)`)
+    console.log(`After appendLinkIfNecessary, ${todosWithLinksMaybe?.length ?? 0} potential items (may include headings or completed)`)
     const tasksByType = todosWithLinksMaybe.length ? getTasksByType(todosWithLinksMaybe, true) : null // puts in object by type of task and enriches with sort info (like priority)
-    // console.log(`After getTasksByType, ${tasksByType?.open.length ?? 0} OPEN items`)
+    console.log(`After getTasksByType, ${tasksByType?.open.length ?? 0} OPEN items`)
     if (deletePreviousCalendarEntries) {
       await deleteCalendarEventsWithTag(timeBlockTag, dateStr)
     }
-    // console.log(`After deleteCalendarEventsWithTag, ${tasksByType?.open.length ?? 0} open items (still)`)
+    console.log(`After deleteCalendarEventsWithTag, ${tasksByType?.open.length ?? 0} open items (still)`)
     if (tasksByType && tasksByType['open'].length) {
       const sortedTodos = tasksByType['open'].length ? sortListBy(tasksByType['open'], '-priority') : []
-      // console.log(`After sortListBy, ${sortedTodos.length} open items`)
+      console.log(`After sortListBy, ${sortedTodos.length} open items`)
       // $FlowIgnore
       await deleteParagraphsContainingString(editorOrNote(note), timeBlockTag) // Delete our timeblocks before scanning note for user-entered timeblocks
       // $FlowIgnore
       await deleteParagraphsContainingString(editorOrNote(note), eventEnteredOnCalTag) // Delete @jgclark timeblocks->calendar breadcrumbs also
       const calendarMapWithEvents = await getPopulatedTimeMapForToday(dateStr, intervalMins, config)
-      // console.log(`After getPopulatedTimeMapForToday, ${calendarMapWithEvents.length} timeMap slots`)
+      console.log(`After getPopulatedTimeMapForToday, ${calendarMapWithEvents.length} timeMap slots`)
       const eventsToTimeblock = getTimeBlockTimesForEvents(calendarMapWithEvents, sortedTodos, config)
-      const { timeBlockTextList, blockList } = eventsToTimeblock
-      // console.log(`After getTimeBlockTimesForEvents, blocks:\n\tblockList.length=${blockList.length} \n\ttimeBlockTextList.length=${timeBlockTextList.length}`)
+      let { timeBlockTextList, blockList } = eventsToTimeblock
+      console.log(`After getTimeBlockTimesForEvents, blocks:\n\tblockList.length=${blockList.length} \n\ttimeBlockTextList.length=${timeBlockTextList.length}`)
+      console.log(`After getTimeBlockTimesForEvents, insertIntoEditor=${insertIntoEditor} createCalendarEntries=${createCalendarEntries}`)
       if (insertIntoEditor || createCalendarEntries) {
-        // console.log(`About to insert ${timeBlockTextList.length} timeblock items into note`)
+        console.log(`After getTimeBlockTimesForEvents, config.createSyncedCopies=${config.createSyncedCopies} todosWithLinksMaybe.length=${todosWithLinksMaybe.length}`)
+        if (config.createSyncedCopies && todosWithLinksMaybe?.length) {
+          console.log(`createSyncedCopies is true, so we will create synced copies of the todosParagraphs: ${todosParagraphs.length} timeblocks`)
+          const sortedParas = getFullParagraphsCorrespondingToSortList(todosParagraphs, sortedTodos)
+          const syncedList = getSyncedCopiesAsList(sortedParas)
+          console.log(`Deleting previous synced list heading and content`)
+          if (syncedList.length && Editor) deleteBlockUnderTitle(Editor, String(config.syncedCopiesTitle), false)
+          console.log(`Inserting synced list content: ${syncedList.length} items`)
+          // $FlowIgnore
+          await insertItemsIntoNote(editorOrNote(note), syncedList, config.syncedCopiesTitle, config.foldSyncedCopiesHeading, config)
+        }
+        console.log(`About to insert ${timeBlockTextList?.length} timeblock items into note`)
         // $FlowIgnore -- Delete any previous timeblocks we created
-        await insertItemsIntoNote(editorOrNote(note), timeBlockTextList, config)
+        await insertItemsIntoNote(editorOrNote(note), timeBlockTextList, config.timeBlockHeading, config.foldTimeBlockHeading, config)
         if (createCalendarEntries) {
-          // console.log(`About to create calendar entries`)
+          console.log(`About to create calendar entries`)
           await writeTimeBlocksToCalendar(getEventsConfig(config), Editor) //using @jgclark's method for now
           if (!insertIntoEditor) {
             // If user didn't want the timeblocks inserted into the editor, then we delete them now that they're in calendar
@@ -329,7 +437,7 @@ export async function createTimeBlocksForTodaysTasks(config: { [key: string]: mi
       }
       return passBackResults ? timeBlockTextList : []
     } else {
-      // console.log('No todos/references marked for >today')
+      console.log('No todos/references marked for >today')
       if (!passBackResults) {
         await showMessage(`No todos/references marked for >today`)
       }
@@ -343,12 +451,18 @@ export async function createTimeBlocksForTodaysTasks(config: { [key: string]: mi
   return []
 }
 
+/**
+ * Insert todos marked >today into the editor
+ * (entry point for /atb)
+ * @param {*} note
+ */
 export async function insertTodosAsTimeblocks(note: TNote): Promise<void> {
-  // console.log(`====== /atb =======\nStarting insertTodosAsTimeblocks`)
+  console.log(`====== /atb =======\nStarting insertTodosAsTimeblocks`)
   if (!editorIsOpenToToday()) await Editor.openNoteByDate(new Date(), false) //open editor to today
   const config = await getConfig()
+  clo(config, 'atb config')
   if (config) {
-    // console.log(`Config found. Calling createTimeBlocksForTodaysTasks`)
+    console.log(`Config found. Calling createTimeBlocksForTodaysTasks`)
     await createTimeBlocksForTodaysTasks(config)
   } else {
     // console.log(`insertTodosAsTimeblocks: stopping after config create`)