+++ helpers/config.js
@@ -26,10 +26,7 @@
  *  console.log(e.message)
  * }
  */
-export function validateConfigProperties(
-  config: { [string]: mixed },
-  validations: { [string]: mixed },
-): { [string]: mixed } {
+export function validateConfigProperties(config: { [string]: mixed }, validations: { [string]: mixed }): { [string]: mixed } {
   let failed = ''
   const propsToValidate = Object.keys(validations)
   if (propsToValidate.length) {
@@ -41,6 +38,7 @@ export function validateConfigProperties(
 
       if (configFieldValue === null || configFieldValue === undefined) {
         if (!isOptional) {
+          console.log(`validateConfigProperties: configFieldValue: ${configFieldValue} for ${v} is null or undefined`)
           failed = `Config required field: "${v}" is missing;\n`
         }
       } else {
@@ -49,8 +47,7 @@ export function validateConfigProperties(
             failed += `Config field: "${v}" (${String(config[v])}) is not the proper type;\n`
           }
         } else {
-          const test =
-            requiredType === 'array' ? Array.isArray(configFieldValue) : typeof configFieldValue === requiredType
+          const test = requiredType === 'array' ? Array.isArray(configFieldValue) : typeof configFieldValue === requiredType
           if (!test) {
             failed += `Config required field: "${v}" is not of type "${String(requiredType)}";\n`
           }